Remove Observed/Deferred/FP status tabs on Workload CVE pages.
