+++ ui/apps/platform/src/Containers/Vulnerabilities/WorkloadCves/Image/ImagePageVulnerabilities.tsx
@@ -11,17 +11,12 @@ import {
     Spinner,
     Split,
     SplitItem,
-    Tab,
-    TabTitleText,
-    Tabs,
-    TabsComponent,
     Text,
     Title,
 } from '@patternfly/react-core';
 import { ExclamationCircleIcon } from '@patternfly/react-icons';
 import { gql, useQuery } from '@apollo/client';
 
-import useURLStringUnion from 'hooks/useURLStringUnion';
 import useURLSearch from 'hooks/useURLSearch';
 import useURLPagination from 'hooks/useURLPagination';
 import useURLSort from 'hooks/useURLSort';
@@ -45,7 +40,6 @@ import {
     getCveStatusScopedQueryString,
     parseQuerySearchFilter,
 } from '../searchUtils';
-import { cveStatusTabValues } from '../types';
 import BySeveritySummaryCard from '../SummaryCards/BySeveritySummaryCard';
 import { imageMetadataContextFragment, ImageMetadataContext } from '../Tables/table.utils';
 import { Resource } from '../components/FilterResourceDropdown';
@@ -76,8 +70,6 @@ export type ImagePageVulnerabilitiesProps = {
 };
 
 function ImagePageVulnerabilities({ imageId }: ImagePageVulnerabilitiesProps) {
-    const [activeTabKey, setActiveTabKey] = useURLStringUnion('cveStatus', cveStatusTabValues);
-
     const { searchFilter } = useURLSearch();
     const querySearchFilter = parseQuerySearchFilter(searchFilter);
     const { page, perPage, setPage, setPerPage } = useURLPagination(20);
@@ -111,7 +103,7 @@ function ImagePageVulnerabilities({ imageId }: ImagePageVulnerabilitiesProps) {
     >(imageVulnerabilitiesQuery, {
         variables: {
             id: imageId,
-            query: getCveStatusScopedQueryString(querySearchFilter, activeTabKey),
+            query: getCveStatusScopedQueryString(querySearchFilter),
             pagination,
         },
     });
@@ -217,39 +209,10 @@ function ImagePageVulnerabilities({ imageId }: ImagePageVulnerabilitiesProps) {
                 className="pf-u-display-flex pf-u-flex-direction-column pf-u-flex-grow-1"
                 component="div"
             >
-                <Tabs
-                    activeKey={activeTabKey}
-                    onSelect={(e, key) => setActiveTabKey(key)}
-                    component={TabsComponent.nav}
-                    mountOnEnter
-                    unmountOnExit
-                    isBox
-                >
-                    <Tab
-                        className="pf-u-display-flex pf-u-flex-direction-column pf-u-flex-grow-1"
-                        eventKey="Observed"
-                        title={<TabTitleText>Observed CVEs</TabTitleText>}
-                    >
-                        <div className="pf-u-px-sm pf-u-background-color-100">
-                            <WorkloadTableToolbar supportedResourceFilters={imageResourceFilters} />
-                        </div>
-                        <div className="pf-u-flex-grow-1 pf-u-background-color-100">
-                            {mainContent}
-                        </div>
-                    </Tab>
-                    <Tab
-                        className="pf-u-display-flex pf-u-flex-direction-column pf-u-flex-grow-1"
-                        eventKey="Deferred"
-                        title={<TabTitleText>Deferrals</TabTitleText>}
-                        isDisabled
-                    />
-                    <Tab
-                        className="pf-u-display-flex pf-u-flex-direction-column pf-u-flex-grow-1"
-                        eventKey="False Positive"
-                        title={<TabTitleText>False positives</TabTitleText>}
-                        isDisabled
-                    />
-                </Tabs>
+                <div className="pf-u-px-sm pf-u-background-color-100">
+                    <WorkloadTableToolbar supportedResourceFilters={imageResourceFilters} />
+                </div>
+                <div className="pf-u-flex-grow-1 pf-u-background-color-100">{mainContent}</div>
             </PageSection>
         </>
     );
