+++ ui/apps/platform/src/Containers/Vulnerabilities/WorkloadCves/Overview/WorkloadCvesOverviewPage.tsx
@@ -7,13 +7,23 @@ import {
     ToolbarItem,
     Flex,
     FlexItem,
+    Card,
+    CardBody,
 } from '@patternfly/react-core';
+import { useQuery } from '@apollo/client';
 
 import useLocalStorage from 'hooks/useLocalStorage';
+import useURLSearch from 'hooks/useURLSearch';
+import useURLStringUnion from 'hooks/useURLStringUnion';
 import PageTitle from 'Components/PageTitle';
-import CveStatusTabNavigation from './CveStatusTabNavigation';
+
+import { VulnMgmtLocalStorage, entityTabValues } from '../types';
+import { parseQuerySearchFilter, getCveStatusScopedQueryString } from '../searchUtils';
 import DefaultFilterModal from '../components/DefaultFilterModal';
-import { VulnMgmtLocalStorage } from '../types';
+import { entityTypeCountsQuery } from '../components/EntityTypeToggleGroup';
+import CVEsTableContainer from './CVEsTableContainer';
+import DeploymentsTableContainer from './DeploymentsTableContainer';
+import ImagesTableContainer from './ImagesTableContainer';
 
 const emptyStorage: VulnMgmtLocalStorage = {
     preferences: {
@@ -25,6 +35,19 @@ const emptyStorage: VulnMgmtLocalStorage = {
 };
 
 function WorkloadCvesOverviewPage() {
+    const { searchFilter } = useURLSearch();
+    const querySearchFilter = parseQuerySearchFilter(searchFilter);
+    const [activeEntityTabKey] = useURLStringUnion('entityTab', entityTabValues);
+
+    const { data: countsData = { imageCount: 0, imageCVECount: 0, deploymentCount: 0 } } = useQuery(
+        entityTypeCountsQuery,
+        {
+            variables: {
+                query: getCveStatusScopedQueryString(querySearchFilter),
+            },
+        }
+    );
+
     const [storedValue, setStoredValue] = useLocalStorage('vulnerabilityManagement', emptyStorage);
 
     function setLocalStorage(values) {
@@ -60,7 +83,30 @@ function WorkloadCvesOverviewPage() {
                 </Flex>
             </PageSection>
             <PageSection padding={{ default: 'noPadding' }}>
-                <CveStatusTabNavigation defaultFilters={storedValue.preferences.defaultFilters} />
+                <PageSection isCenterAligned>
+                    <Card>
+                        <CardBody>
+                            {activeEntityTabKey === 'CVE' && (
+                                <CVEsTableContainer
+                                    defaultFilters={storedValue.preferences.defaultFilters}
+                                    countsData={countsData}
+                                />
+                            )}
+                            {activeEntityTabKey === 'Image' && (
+                                <ImagesTableContainer
+                                    defaultFilters={storedValue.preferences.defaultFilters}
+                                    countsData={countsData}
+                                />
+                            )}
+                            {activeEntityTabKey === 'Deployment' && (
+                                <DeploymentsTableContainer
+                                    defaultFilters={storedValue.preferences.defaultFilters}
+                                    countsData={countsData}
+                                />
+                            )}
+                        </CardBody>
+                    </Card>
+                </PageSection>
             </PageSection>
         </>
     );
